+++ starlite/routes/__init__.py
@@ -1,0 +1,6 @@
+from .asgi import ASGIRoute
+from .base import BaseRoute
+from .http import HTTPRoute
+from .websocket import WebSocketRoute
+
+__all__ = ["BaseRoute", "ASGIRoute", "WebSocketRoute", "HTTPRoute"]
